Use feed names provided by the user as the sender of emails
Instead of using the feed title, use the feed names that the user
provided in the configuration file as the name of the sender of the
emails for that feed.

This allows the user to control it (some feeds have too long titles and
it is annoying to have them as the senders' names).

This only affects the sender but not how the data is stored in the
database so this does not introduce any regression (no migration is
required)