Check for non-zero return code
After the call, check the exit code and it is non-zero, raise a
CalledProcessError like the subprocess.run function does.

On failure print the standard error of the failed program. When we drop
support of 2.7 and 3.2 and lower, we can add the standard error as part
of the CalledProcessError and handle it later.