Feed update time is max(newest post time, feed update time)
Some RSS sources (like Youtube) don't update the feed update time
(feed.feed.updated). It is a valid date but it is not updated when a new
post is added.

To overcome this we get the update time of the newest post and compare
it with the feed update time. We use the newest of both (max).