Fetch and parse feeds in parallel using threads
Setup a thread pool of 4 threads (by default) to fetch and parse the
feeds in parallel.

In my personal setup I have 23 feeds and it takes to fetch and parse
them in sequence around 24 secs. With the thread pool of 4 threads the
time was reduced to 8 secs.

The speedup will vary but the experiment shown that `feedparser` is IO
bound so the use of threads will improve the overall performance.